Extract deleteWasSpace function.
