+++ app/screens/WAS/WasScreen.tsx
@@ -12,13 +12,10 @@ import { navigationRef } from '../../navigation/navigationRef'
 import { Ed25519VerificationKey2020 } from '@digitalcredentials/ed25519-verification-key-2020'
 
 import AsyncStorage from '@react-native-async-storage/async-storage'
-// import FileReader from 'react-native-filereader';
-import { v4 as uuidv4 } from 'uuid'
+// import FileReader from 'react-native-filereader'
 import { WAS } from '../../../app.config'
 import { useThemeContext } from '../../hooks'
-import { removeWasPublicLink } from '../../lib/removeWasPublicLink'
-import { getRootSigner } from '../../lib/getRootSigner'
-import { getStorageClient, provisionWasSpace } from '../../lib/walletAttachedStorage'
+import { deleteWasSpace, provisionWasSpace } from '../../lib/walletAttachedStorage'
 
 const WASScreen = () => {
   const { theme } = useThemeContext()
@@ -41,40 +38,7 @@ const WASScreen = () => {
         throw new Error('No connection details found')
       }
 
-      // Get the stored signer
-      const signer = await getRootSigner()
-      const storage = getStorageClient()
-      const space = storage.space({
-        signer,
-        id: connectionDetails.spaceId as `urn:uuid:${string}`
-      })
-
-      // Delete the space
-      const response = await space.delete({
-        signer
-      })
-
-      if (!response.ok) {
-        throw new Error(`Failed to delete space. Status: ${response.status}`)
-      }
-
-      const mapData = await AsyncStorage.getItem('map')
-      if (mapData) {
-        const map = JSON.parse(mapData)
-        // Get all keys that start with 'publiclinks_'
-        const publicLinkKeys = Object.keys(map)
-          .filter((key) => key.startsWith('publiclinks_'))
-          .map((key) => key.replace('publiclinks_', ''))
-
-        // Process each public link
-        for (const key of publicLinkKeys) {
-          await removeWasPublicLink(key, map)
-        }
-      }
-
-      // Clear stored items
-      await AsyncStorage.removeItem(WAS.KEYS.SIGNER_KEYPAIR)
-      await AsyncStorage.removeItem(WAS.KEYS.SPACE_ID)
+      await deleteWasSpace({ spaceId: connectionDetails.spaceId });
 
       setStatus('success')
       setMessage('Space successfully deleted')
